Module canvas::canvas_paint_thread

Structs

CanvasPaintState

CanvasPaintThread

Traits

PointToi32

RectToi32

SizeToi32

ToAzFloat

Functions

crop_image
Used by drawImage to get rid of the extra pixels of the image data that won't be copied to the canvas image_data: Color pixel data of the image image_size: Image dimensions crop_rect: It determines the area of the image we want to keep

is_zero_size_gradient

write_image
It writes an image to the destination target draw_target: the destination target where the image_data will be copied image_data: Pixel information of the image to be written. It takes RGBA8 image_size: The size of the image to be written dest_rect: Area of the destination target where the pixels will be copied smoothing_enabled: It determines if smoothing is applied to the image result