canvas::vello_cpu_backend

Struct VelloCPUDrawTarget

pub(crate) struct VelloCPUDrawTarget {
    ctx: RenderContext,
    pixmap: Pixmap,
    clips: Vec<Path>,
    state: State,
}

Fields
ctx: RenderContext

Because this is stateful context caller cannot assume anything about transform, paint, stroke, so it should provide it’s own used by each command but it can assume paint_transform to be identity and fill rule to be peniko::Fill::NonZero

This is because paint_transform is rarely set, so it’s cheaper to always reset it after use.

pixmap: Pixmapclips: Vec<Path>state: State
Implementations

impl VelloCPUDrawTarget

fn with_composition( &mut self, composition_operation: CompositionOrBlending, f: impl FnOnce(&mut Self), )

fn ignore_clips(&mut self, f: impl FnOnce(&mut Self))

fn ensure_drawing(&mut self)

fn pixmap(&mut self) -> &[u8]

fn size(&self) -> Size2D<u32>

fn is_viewport_cleared( &mut self, rect: &Rect<f32>, transform: Transform2D<f64>, ) -> bool