Module canvas::webgl_thread

Structs

FramebufferRebindingFlags 🔒

FramebufferRebindingInfo 🔒

GLContextData 🔒

GLState

WebGLContextInfo 🔒

Helper struct to store cached WebGLContext information.

WebGLImpl

WebGL Commands Implementation

WebGLThread 🔒

A WebGLThread manages the life cycle and message multiplexing of a set of WebGLContexts living in the same thread.

WebGLThreadInit 🔒

The data required to initialize an instance of the WebGLThread type.

WebXRBridge 🔒

Bridge between WebGL and WebXR

WebXRBridgeContexts 🔒

WebXRBridgeGrandManager 🔒

WebXRBridgeInit 🔒

WebXRBridgeManager 🔒

Constants

ANGLE_NAME_PREFIX 🔒

ANGLE adds a _u prefix to variable names:

SAFE_VIEWPORT_DIMS 🔒

Traits

SurfmanContextAttributeFlagsConvert 🔒

ToSurfmanVersion 🔒

Functions

clamp_viewport 🔒

current_wr_image_buffer_kind 🔒

flip_pixels_y 🔒

Flips the pixels in the Vec on the Y axis.

from_name_in_compiled_shader 🔒

image_to_tex_image_data 🔒

Translates an image in rgba8 (red in the first byte) format to the format that was requested of TexImage.

map_dot_separated 🔒

premultiply_inplace 🔒

prepare_pixels 🔒

to_name_in_compiled_shader 🔒

unmultiply_inplace 🔒