Struct canvas::canvas_data::CanvasData

pub struct CanvasData<'a> {
    backend: Box<dyn Backend>,
    drawtarget: Box<dyn GenericDrawTarget>,
    path_state: Option<PathState>,
    state: CanvasPaintState<'a>,
    saved_states: Vec<CanvasPaintState<'a>>,
    compositor_api: CrossProcessCompositorApi,
    image_key: Option<ImageKey>,
    old_image_key: Option<ImageKey>,
    very_old_image_key: Option<ImageKey>,
    font_context: Arc<FontContext>,
}

Fields
backend: Box<dyn Backend>drawtarget: Box<dyn GenericDrawTarget>path_state: Option<PathState>state: CanvasPaintState<'a>saved_states: Vec<CanvasPaintState<'a>>compositor_api: CrossProcessCompositorApiimage_key: Option<ImageKey>old_image_key: Option<ImageKey>

An old webrender image key that can be deleted when the next epoch ends.

very_old_image_key: Option<ImageKey>

An old webrender image key that can be deleted when the current epoch ends.

font_context: Arc<FontContext>
Implementations

impl<'a> CanvasData<'a>

pub fn new( size: Size2D<u64>, compositor_api: CrossProcessCompositorApi, antialias: AntialiasMode, font_context: Arc<FontContext>, ) -> CanvasData<'a>

pub fn draw_image( &mut self, image_data: &[u8], image_size: Size2D<f64>, dest_rect: Rect<f64>, source_rect: Rect<f64>, smoothing_enabled: bool, premultiply: bool, )

pub fn save_context_state(&mut self)

pub fn restore_context_state(&mut self)

pub fn fill_text_with_size( &mut self, text: String, x: f64, y: f64, max_width: Option<f64>, is_rtl: bool, size: f64, )

pub fn fill_text( &mut self, text: String, x: f64, y: f64, max_width: Option<f64>, is_rtl: bool, )

https://html.spec.whatwg.org/multipage/#text-preparation-algorithm

pub fn measure_text(&mut self, text: String) -> TextMetrics

https://html.spec.whatwg.org/multipage/#text-preparation-algorithm https://html.spec.whatwg.org/multipage/#dom-context-2d-measuretext

fn build_unshaped_text_runs<'b>( &self, text: &'b str, font_group: &mut FontGroup, ) -> Vec<UnshapedTextRun<'b>>

fn find_anchor_point_for_line_of_text( &self, x: f32, y: f32, metrics: &FontMetrics, width: f32, is_rtl: bool, ) -> Point2D<f32>

Find the anchor_point for the given parameters of a line of text. See https://html.spec.whatwg.org/multipage/#text-preparation-algorithm.

pub fn fill_rect(&mut self, rect: &Rect<f32>)

pub fn clear_rect(&mut self, rect: &Rect<f32>)

pub fn stroke_rect(&mut self, rect: &Rect<f32>)

pub fn begin_path(&mut self)

pub fn close_path(&mut self)

fn ensure_path(&mut self)

fn path(&self) -> &Path

pub fn fill(&mut self)

pub fn stroke(&mut self)

pub fn clip(&mut self)

pub fn is_point_in_path( &mut self, x: f64, y: f64, _fill_rule: FillRule, chan: IpcSender<bool>, )

pub fn move_to(&mut self, point: &Point2D<f32>)

pub fn line_to(&mut self, point: &Point2D<f32>)

fn path_builder(&mut self) -> PathBuilderRef<'_>

pub fn rect(&mut self, rect: &Rect<f32>)

pub fn quadratic_curve_to(&mut self, cp: &Point2D<f32>, endpoint: &Point2D<f32>)

pub fn bezier_curve_to( &mut self, cp1: &Point2D<f32>, cp2: &Point2D<f32>, endpoint: &Point2D<f32>, )

pub fn arc( &mut self, center: &Point2D<f32>, radius: f32, start_angle: f32, end_angle: f32, ccw: bool, )

pub fn arc_to(&mut self, cp1: &Point2D<f32>, cp2: &Point2D<f32>, radius: f32)

pub fn ellipse( &mut self, center: &Point2D<f32>, radius_x: f32, radius_y: f32, rotation_angle: f32, start_angle: f32, end_angle: f32, ccw: bool, )

pub fn set_fill_style(&mut self, style: FillOrStrokeStyle)

pub fn set_stroke_style(&mut self, style: FillOrStrokeStyle)

pub fn set_line_width(&mut self, width: f32)

pub fn set_line_cap(&mut self, cap: LineCapStyle)

pub fn set_line_join(&mut self, join: LineJoinStyle)

pub fn set_miter_limit(&mut self, limit: f32)

pub fn get_transform(&self) -> Transform2D<f32>

pub fn set_transform(&mut self, transform: &Transform2D<f32>)

pub fn set_global_alpha(&mut self, alpha: f32)

pub fn set_global_composition(&mut self, op: CompositionOrBlending)

pub fn recreate(&mut self, size: Option<Size2D<u64>>)

pub fn send_pixels(&mut self, chan: IpcSender<IpcSharedMemory>)

pub fn send_data(&mut self, chan: IpcSender<CanvasImageData>)

pub fn put_image_data(&mut self, imagedata: Vec<u8>, rect: Rect<u64>)

pub fn set_shadow_offset_x(&mut self, value: f64)

pub fn set_shadow_offset_y(&mut self, value: f64)

pub fn set_shadow_blur(&mut self, value: f64)

pub fn set_shadow_color(&mut self, value: AbsoluteColor)

pub fn set_font(&mut self, font_style: FontStyleStruct)

pub fn set_text_align(&mut self, text_align: TextAlign)

pub fn set_text_baseline(&mut self, text_baseline: TextBaseline)

fn need_to_draw_shadow(&self) -> bool

fn create_draw_target_for_shadow( &self, source_rect: &Rect<f32>, ) -> Box<dyn GenericDrawTarget>

fn draw_with_shadow<F>(&self, rect: &Rect<f32>, draw_shadow_source: F)
where F: FnOnce(&mut dyn GenericDrawTarget),

pub fn read_pixels( &self, read_rect: Rect<u64>, canvas_size: Size2D<u64>, ) -> Vec<u8>

It reads image data from the canvas canvas_size: The size of the canvas we’re reading from read_rect: The area of the canvas we want to read from