Module canvas::webgl_thread

Structs

FramebufferRebindingFlags ğŸ”’

FramebufferRebindingInfo ğŸ”’

GLContextData ğŸ”’

GLState

WebGLContextInfo ğŸ”’

Helper struct to store cached WebGLContext information.

WebGLImpl

WebGL Commands Implementation

WebGLThread ğŸ”’

A WebGLThread manages the life cycle and message multiplexing of a set of WebGLContexts living in the same thread.

WebGLThreadInit ğŸ”’

The data required to initialize an instance of the WebGLThread type.

WebXRBridge ğŸ”’

Bridge between WebGL and WebXR

WebXRBridgeContexts ğŸ”’

WebXRBridgeGrandManager ğŸ”’

WebXRBridgeInit ğŸ”’

WebXRBridgeManager ğŸ”’

Constants

ANGLE_NAME_PREFIX ğŸ”’

ANGLE adds a _u prefix to variable names:

SAFE_VIEWPORT_DIMS ğŸ”’

Traits

SurfmanContextAttributeFlagsConvert ğŸ”’

ToSurfmanVersion ğŸ”’

Functions

clamp_viewport ğŸ”’

current_wr_image_buffer_kind ğŸ”’

flip_pixels_y ğŸ”’

Flips the pixels in the Vec on the Y axis.

from_name_in_compiled_shader ğŸ”’

image_to_tex_image_data ğŸ”’

Translates an image in rgba8 (red in the first byte) format to the format that was requested of TexImage.

map_dot_separated ğŸ”’

premultiply_inplace ğŸ”’

prepare_pixels ğŸ”’

to_name_in_compiled_shader ğŸ”’