Struct canvas::webgl_thread::WebGLThread

pub(crate) struct WebGLThread {

    device: Device,
    webrender_api: RenderApi,
    webrender_doc: DocumentId,
    contexts: FnvHashMap<WebGLContextId, GLContextData>,
    cached_context_info: FnvHashMap<WebGLContextId, WebGLContextInfo>,
    bound_context_id: Option<WebGLContextId>,
    dom_outputs: FnvHashMap<PipelineId, DOMToTextureData>,
    external_images: Arc<Mutex<WebrenderExternalImageRegistry>>,
    receiver: Receiver<WebGLMsg>,
    sender: WebGLSender<WebGLMsg>,
    webrender_swap_chains: SwapChains<WebGLContextId, Device>,
    api_type: GlType,
    pub webxr_bridge: WebXRBridge,

}

A WebGLThread manages the life cycle and message multiplexing of a set of WebGLContexts living in the same thread.

Fields
device: Device

The GPU device.

webrender_api: RenderApi

Channel used to generate/update or delete webrender_api::ImageKeys.

webrender_doc: DocumentIdcontexts: FnvHashMap<WebGLContextId, GLContextData>

Map of live WebGLContexts.

cached_context_info: FnvHashMap<WebGLContextId, WebGLContextInfo>

Cached information for WebGLContexts.

bound_context_id: Option<WebGLContextId>

Current bound context.

dom_outputs: FnvHashMap<PipelineId, DOMToTextureData>

Texture ids and sizes used in DOM to texture outputs.

external_images: Arc<Mutex<WebrenderExternalImageRegistry>>

List of registered webrender external images. We use it to get an unique ID for new WebGLContexts.

receiver: Receiver<WebGLMsg>

The receiver that will be used for processing WebGL messages.

sender: WebGLSender<WebGLMsg>

The receiver that should be used to send WebGL messages for processing.

webrender_swap_chains: SwapChains<WebGLContextId, Device>

The swap chains used by webrender

api_type: GlType

Whether this context is a GL or GLES context.

webxr_bridge: WebXRBridge

The bridge to WebXR

Implementations

impl WebGLThread

pub(crate) fn new(_: WebGLThreadInit) -> Self

Create a new instance of WebGLThread.

pub(crate) fn run_on_own_thread(init: WebGLThreadInit)

Perform all initialization required to run an instance of WebGLThread in parallel on its own dedicated thread.

fn process(&mut self)

fn handle_msg(&mut self, msg: WebGLMsg, webgl_chan: &WebGLChan) -> bool

Handles a generic WebGLMsg message

fn handle_webxr_command(&mut self, command: WebXRCommand)

Handles a WebXR message

fn handle_webgl_command(
&mut self,
context_id: WebGLContextId,
command: WebGLCommand,
backtrace: WebGLCommandBacktrace
)

Handles a WebGLCommand for a specific WebGLContext

fn create_webgl_context(
&mut self,
webgl_version: WebGLVersion,
requested_size: Size2D<u32>,
attributes: GLContextAttributes
) -> Result<(WebGLContextId, GLLimits), String>

Creates a new WebGLContext

fn resize_webgl_context(
&mut self,
context_id: WebGLContextId,
requested_size: Size2D<u32>
) -> Result<(), String>

Resizes a WebGLContext

fn remove_webgl_context(&mut self, context_id: WebGLContextId)

Removes a WebGLContext and releases attached resources.

fn handle_swap_buffers(
&mut self,
context_ids: Vec<WebGLContextId>,
completed_sender: WebGLSender<u64>,
_sent_time: u64
)

fn surface_access(&self) -> SurfaceAccess

Which access mode to use

fn handle_dom_to_texture(&mut self, command: DOMToTextureCommand)

pub(crate) fn handle_dom_to_texture_lock(
&mut self,
pipeline_id: PipelineId,
gl_sync: usize
) -> Option<(u32, Size2D<i32>)>

fn make_current_if_needed<'a>(
device: &Device,
context_id: WebGLContextId,
contexts: &'a FnvHashMap<WebGLContextId, GLContextData>,
bound_id: &mut Option<WebGLContextId>
) -> Option<&'a GLContextData>

Gets a reference to a Context for a given WebGLContextId and makes it current if required.

fn make_current_if_needed_mut<'a>(
device: &Device,
context_id: WebGLContextId,
contexts: &'a mut FnvHashMap<WebGLContextId, GLContextData>,
bound_id: &mut Option<WebGLContextId>
) -> Option<&'a mut GLContextData>

Gets a mutable reference to a GLContextWrapper for a WebGLContextId and makes it current if required.

fn create_wr_external_image(
webrender_api: &mut RenderApi,
webrender_doc: DocumentId,
size: Size2D<i32>,
alpha: bool,
context_id: WebGLContextId,
target: TextureTarget
) -> ImageKey

Creates a webrender_api::ImageKey that uses shared textures.

fn update_wr_external_image(
webrender_api: &mut RenderApi,
webrender_doc: DocumentId,
size: Size2D<i32>,
alpha: bool,
context_id: WebGLContextId,
image_key: ImageKey,
target: TextureTarget
)

Updates a webrender_api::ImageKey that uses shared textures.

fn image_descriptor(size: Size2D<i32>, alpha: bool) -> ImageDescriptor

Helper function to create a webrender_api::ImageDescriptor.

fn external_image_data(
context_id: WebGLContextId,
target: TextureTarget
) -> ImageData

Helper function to create a webrender_api::ImageData::External instance.

fn get_glsl_version(gl: &Gl) -> WebGLSLVersion

Gets the GLSL Version supported by a GLContext.